+++ test/data/recipe_loader_test.dart
@@ -1,3 +1,4 @@
+@Tags(['legacy'])
 import 'dart:convert';
 
 import 'package:flutter/services.dart';